helper/requirements.txt
altgraph==0.17.3
anyio==3.6.2
autopep8==2.0.0
bidict==0.22.0
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
fastapi==0.86.0
frida==15.2.2
frida-tools==11.0.0
h11==0.14.0
hexdump==3.3
idna==3.4
macholib==1.16.2
netifaces==0.10.6
prompt-toolkit==3.0.32
pycodestyle==2.9.1
pydantic==1.10.2
Pygments==2.13.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.13
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
